Set up multiple bots
