+++ .gitignore
@@ -160,3 +160,4 @@ cython_debug/
 #.idea/
 
 .serverless
+training_storage
