--- data/2023-updates.txt
@@ -1,3 +1,0 @@
-In New Jersey, the maximum weekly benefit amount for 2023 is $830.
-
-The phone number to certify (claim) benefits is (732) 937-4525.
