--- requirements.txt
@@ -1,2 +1,0 @@
-llama_index==0.7.2
-openai==0.27.8
