+++ src/anchor/data/urls.txt
@@ -1,0 +1,10 @@
+https://nj.gov/treasury/taxation/anchor/index.shtml
+https://nj.gov/treasury/taxation/anchor/home.shtml
+https://nj.gov/treasury/taxation/anchor/tenant.shtml
+https://nj.gov/treasury/taxation/anchor/home-paperapp.shtml
+https://nj.gov/treasury/taxation/pdf/anchor/anchor-h.pdf
+https://nj.gov/treasury/taxation/pdf/anchor/anchor-t.pdf
+https://nj.gov/treasury/taxation/anchor/home-linebyline.shtml
+https://nj.gov/treasury/taxation/anchor/home-faq.shtml
+https://nj.gov/treasury/taxation/anchor/tenant-calculated.shtml
+https://nj.gov/treasury/taxation/anchor/tenant-faq.shtml
