+++ src/anchor/requirements.txt
@@ -1,0 +1,2 @@
+llama_index==0.7.2
+openai==0.27.8
