+++ src/anchor/test.py
@@ -1,0 +1,9 @@
+from handler import ask
+
+event = {
+  "queryStringParameters": {
+    "question": "what is this program"
+  }
+}
+
+ask(event=event, context={})
