--- training_storage/graph_store.json
@@ -1,1 +1,0 @@
-{"graph_dict": {}}
